+++ src/main/java/com/mraof/minestuck/alchemy/GristHelper.java
@@ -10,7 +10,6 @@ import net.minecraftforge.common.MinecraftForge;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 
 public class GristHelper
 {
@@ -69,28 +68,4 @@ public class GristHelper
 		return event.getNewDrops();
 		
 	}
-	
-	//TODO figure out how best to check cache capacity client-side
-	public static boolean canAfford(GristSet base, GristSet cost)
-	{
-		if(base == null || cost == null)
-		{
-			return false;
-		}
-		Map<GristType, Long> reqs = cost.getMap();
-		if(reqs != null)
-		{
-			for(Entry<GristType, Long> pairs : reqs.entrySet())
-			{
-				GristType type = pairs.getKey();
-				long need = pairs.getValue();
-				long have = base.getGrist(type);
-				
-				if(need > have) return false;
-			}
-			return true;
-		}
-		return false;
-	}
-	
 }