+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Fixed a crash with TerraBlender
 - Fixed grist cache view in editmode using the limit of the wrong player
+- Cache limit will now be accounted for more often client-side
 
 ### Contributors for this release